requirements.txt
requests
beautifulsoup4
lxml
discord.py
python-dotenv